remove simphotovoltaic fast completely
